requirements.txt
numpy 
pandas 
python-dotenv 
openai          
openpyxl
